star_identification/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
Pillow>=9.4.0
tqdm>=4.65.0
wandb>=0.15.0
matplotlib>=3.5.0
seaborn>=0.12.0
albumentations>=1.3.0